Fixed changes not being explicit to certain crops
Using Netcode with very little understanding and documentation, I managed to figure out how to check if a crop is a specific type of crop, thus giving me the option to use logic to only change the crops I wanted. This also gave me the ability to make an anti mod, which is also in my repos.